Project page (#84)
* feat: Add bunnet dependency and update openai version

The code changes include adding the `bunnet` dependency to the `requirements.txt` file and updating the `openai` dependency to version 1.40.2. These changes are necessary to support the new functionality introduced in the code.

Recent user commits:
- popup on upload
- multi file upload
- Add a celery task queue to backend (#82)
- Update README.md
- Merge pull request #81 from Watts-Lab/markwhiting-patch-1
- Update README.md
- assign docker command to task definition
- clean up docker compose for development env
- Login window (#76)
- Update README.md
- upload button

Recent repository commits:
- popup on upload
- multi file upload
- Add a celery task queue to backend (#82)
- Update README.md
- Merge pull request #81 from Watts-Lab/markwhiting-patch-1
- Update README.md
- assign docker command to task definition
- clean up docker compose for development env
- Login window (#76)
- Update README.md
- upload button

* refactor: Rename Papers model to Paper

The code changes rename the `Papers` model to `Paper` in the `papers.py` file. This change improves code clarity and follows the singular naming convention for models.

* project basic working stage

* fix project consitansy + pagination